# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Backend",
        "description": "Admin Backend Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Admin Backend"
        }
    ],
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "operationId": " All Subjects",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['posts']"
                    }
                }
            }
        },
...
        "/api/streamer/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "operationId": "Subjects of streamer",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['subjects']"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}